requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.2
essentia==2.1b6.dev1389
idna==3.10
numpy==2.3.2
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.3.0
requests==2.32.4
six==1.17.0
spotipy==2.25.1
urllib3==2.5.0
ytmusicapi==1.11.0
